Fix bugs causing wrong pre-synaptic innervation
The synapses is sorted so that those innervating the same post synaptic cite
(either neuron or synapse) are adjacent, and so the post-synaptic neurons could
access synaptic data coercingly. However, this setting messes up the id of
synapses, and causes problem to synapses whose post-synaptic cite is another
synapse. We need to remap the synapse id so that the configuration is setup
correctly.